silent action #228

+++ tests-ng/actions.sh
@@ -65,6 +65,7 @@ actions:
     postaction: echo 'post' > ${tmpa}/post
     postaction2: echo 'post2' > ${tmpa}/post2
   nakedaction: echo 'naked' > ${tmpa}/naked
+  _silentaction: echo 'silent'
 config:
   backup: true
   create: true
@@ -79,6 +80,7 @@ dotfiles:
       - nakedaction
       - preaction2
       - postaction2
+      - _silentaction
 profiles:
   p1:
     dotfiles:
@@ -90,7 +92,7 @@ _EOF
 echo "test" > ${tmps}/dotfiles/abc
 
 # install
-cd ${ddpath} | ${bin} install -f -c ${cfg} -p p1 -V
+cd ${ddpath} | ${bin} install -f -c ${cfg} -p p1 -V 2>&1 | tee ${tmpa}/log
 
 # checks
 [ ! -e ${tmpa}/pre ] && exit 1
@@ -102,7 +104,11 @@ grep naked ${tmpa}/naked >/dev/null
 [ ! -e ${tmpa}/pre2 ] && exit 1
 grep pre2 ${tmpa}/pre2 >/dev/null
 [ ! -e ${tmpa}/post2 ] && exit 1
-grep post2 ${tmpa}/post2 >/dev/null
+grep post ${tmpa}/post2 >/dev/null
+[ ! -e ${tmpa}/log ] && exit 1
+grep "executing \"echo 'naked' > ${tmpa}/naked" ${tmpa}/log >/dev/null
+grep "executing \"echo 'silent'" ${tmpa}/log >/dev/null && false
+grep "executing silent action \"_silentaction\"" ${tmpa}/log >/dev/null
 
 ## CLEANING
 rm -rf ${tmps} ${tmpd} ${tmpa}
